+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added logic to show default or `date` granularity initially.
+* Clear existing data when granularity is changed.
+* Set Onset date to first day of `month` or `year` as per chosen granularity.
+* Calculation for leap year corrected.
 
 4.14.0 - (January 28, 2020)
 ----------